+++ docs/changelog.md
@@ -6,6 +6,8 @@ Numbers in brackets show the issue number in https://github.com/wp-document-revi
 
 * Extend Validation structure process to identify inaccessible document files and potentially delete them.
 * Provide a filter 'document_validate_orphans' to control the list of attachments considered inaccessible.
+* Review the revision log metabox to only permit restore of revisions that link to a different document file.
+* Fix the update to the age of revisions being displayed in the revision log.
 * Review REST processing to further protect attachment details.
 * Make use of a (temporary) postmeta value to keep track of the current document attachment.
 * Refactor to include class variables in trait files if only used there.